+++ tests/rate_limiter.test.ts
@@ -10,7 +10,6 @@ import {
   expect,
 } from 'vitest';
 
-import { v4 } from 'uuid';
 import {
   FlowProducer,
   Queue,
@@ -20,7 +19,7 @@ import {
   UnrecoverableError,
   Job,
 } from '../src/classes';
-import { delay, removeAllQueueData } from '../src/utils';
+import { delay, randomUUID, removeAllQueueData } from '../src/utils';
 
 describe('Rate Limiter', () => {
   const redisHost = process.env.REDIS_HOST || 'localhost';
@@ -35,7 +34,7 @@ describe('Rate Limiter', () => {
   });
 
   beforeEach(async () => {
-    queueName = `test-${v4()}`;
+    queueName = `test-${randomUUID()}`;
     queue = new Queue(queueName, { connection, prefix });
     queueEvents = new QueueEvents(queueName, { connection, prefix });
     await queueEvents.waitUntilReady();
@@ -422,7 +421,7 @@ describe('Rate Limiter', () => {
   describe('when using flows', () => {
     it('should obey the rate limit per queue', { timeout: 20000 }, async () => {
       const name = 'child-job';
-      const parentQueueName = `parent-queue-${v4()}`;
+      const parentQueueName = `parent-queue-${randomUUID()}`;
       const parentQueueEvents = new QueueEvents(parentQueueName, {
         connection,
         prefix,